refactor requests module
