+++ src/modules/requests/requests.interface.ts
@@ -13,3 +13,18 @@ export interface ChangeStatusBody {
 export interface UpdateRequestItemBody {
   available_item_name: string;
 }
+
+export interface ResponseInterface {
+  id: string;
+  email: string;
+  username: string;
+  division: string;
+  phone_number: string;
+  request_type: number;
+  item_name: string;
+  purpose: string;
+  priority: number;
+  status: number;
+  created_at: Date;
+  updated_at: Date;
+}
