+++ src/modules/requests/requests.repository.ts
@@ -1,7 +1,7 @@
 import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Request } from '../../../entities/request.entity';
-import { UserAccess } from '../../../common/interfaces/keycloak-user.interface';
+import { Request } from '../../entities/request.entity';
+import { UserAccess } from '../../common/interfaces/keycloak-user.interface';
 import { Pagination } from 'src/common/interfaces/pagination.interface';
 export class RequestsRepository extends Repository<Request> {
   constructor(
